+++ packages/story-editor/src/app/highlights/context.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,4 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default as useQuickActions, MediaPicker } from './useQuickActions';
+
+/**
+ * External dependencies
+ */
+import { createContext } from '@googleforcreators/react';
+/**
+ * Internal dependencies
+ */
+import type { HighlightsState } from '../../types/highlightsProvider';
+
+export default createContext<HighlightsState>({});
