+++ packages/story-editor/src/app/highlights/styles.ts
