+++ packages/story-editor/tsconfig.json
@@ -4,5 +4,9 @@
     "rootDir": "src",
     "declarationDir": "dist-types"
   },
-  "include": ["src/app/config/*", "src/types"]
+  "include": [
+    "src/app/config/*",
+    "src/types",
+    "src/app/highlights/*"
+  ]
 }
